removed key missing from list items -console warns from result
