+++ frontend/components/categoryResult.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react'
+import React from 'react'
 import styled from 'styled-components'
 import CategoryResultChart from './categoryResultChart'
 
@@ -44,45 +44,32 @@ const CategoryResult = ({
   category,
   description,
   index,
-}) => {
-  const [renderMobileLayout, setRenderMobileLayout] = useState(false)
-
-  useEffect(() => {
-    const handleResize = () => {
-      window.innerWidth <= 800 ?
-        setRenderMobileLayout(true) :
-        setRenderMobileLayout(false)
-    }
-
-    window.addEventListener('resize', handleResize)
-  }, [window.innerWidth])
-
+  renderMobileLayout
+}) => (
   // if mobile should be rendered, the toggle disables the desktop feature 
   // that puts speedometer on alternating sides of the text
   // -> speedometer goes above text in every categoryresult
-  return (
-    <CategoryResultContainer>
-      {(index % 2 === 0 && !renderMobileLayout) && 
-        <CategoryText>
-          <CategoryTitle>
-            {category} {userResult} / {maxResult}
-          </CategoryTitle>
-          {description}
-        </CategoryText>
-      }
-      <CategoryImage>
-        <CategoryResultChart userResult={userResult} maxResult={maxResult} />
-      </CategoryImage>
-      {(index % 2 !== 0 || renderMobileLayout) && 
-        <CategoryText>
-          <CategoryTitle>
-            {category} {userResult} / {maxResult}
-          </CategoryTitle>
-          {description}
-        </CategoryText>
-      }
-    </CategoryResultContainer>
-  )
-}
+  <CategoryResultContainer>
+    {(index % 2 === 0 && !renderMobileLayout) && 
+      <CategoryText>
+        <CategoryTitle>
+          {category} {userResult} / {maxResult}
+        </CategoryTitle>
+        {description}
+      </CategoryText>
+    }
+    <CategoryImage>
+      <CategoryResultChart userResult={userResult} maxResult={maxResult} />
+    </CategoryImage>
+    {(index % 2 !== 0 || renderMobileLayout) && 
+      <CategoryText>
+        <CategoryTitle>
+          {category} {userResult} / {maxResult}
+        </CategoryTitle>
+        {description}
+      </CategoryText>
+    }
+  </CategoryResultContainer>
+)
 
 export default CategoryResult
